add mlr program list & move to separate dir
